requirements.txt
streamlit>=1.39.0
openai>=1.54.0
pypdf>=5.1.0
python-docx>=1.1.2
python-dotenv>=1.0.1
